refactor: do not type check the first param of a dapi fn before extracting it
Instead of only extracting the first param of a dapi fn if its type is DEPENDENCIES, just assume it
is and always extract the first parameter. We sacrificing type safety with flexibility. Not checking
the dependecy type will allow library users to define deps of only the needed types on their dapi
fns.